refactor: DriverTrait and KeywordsTrait are removed
